moved compare docs method to utils, so that it can used on other tests
